I carried out the following for my project
1. Import a table of your choice into MySQL Workbench and carry out the
following on different tabs.
a. Show distinct values on any of the fields on your table.
b. Create a function on the table.
c. Create Insert, Delete and Update triggers on your table.
d. Create stored procedures.
e. Add a new field to the table.
f. Change the name of a field on your table.
g. Create a new user in your database
h. Grant the new user select, insert and alter privileges.
i. Revoke all the privileges granted to the new user
2. a. Write SQL statements to backup and restore your database.
b. What is the difference between CHAR and VARCHAR datatypes?
c. Create an ERD with 6 related tables.
 d. What is the difference between SQL and MySQL?
 e. Why are triggers created in a database?
f. What is the difference between a Primary Key and a Foreign Key?